+++ src/AboutDialog.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <Windows.h>
+
+class AboutDialog {
+public:
+    static void Show(HWND owner);
+};