+++ src/resource.h
@@ -2,3 +2,12 @@
 
 #define IDI_ICON_CONNECTED      101
 #define IDI_ICON_DISCONNECTED   102
+
+#define IDB_LOGO                201
+
+#define IDD_ABOUT               300
+#define IDC_ABOUT_LOGO          301
+#define IDC_ABOUT_VERSION       302
+#define IDC_ABOUT_MAINTAINER    303
+#define IDC_ABOUT_GITHUB        304
+#define IDC_ABOUT_ISSUES        305